class Konto aus packet 8 entendet mit class Klassen aus packet 3. Ein wichtigen Konstruktor in class Konto hinzu gefügt und paket 3 umgeschrievben sodass class Konto nutzbar ist.
