+++ 00_bwl/bwl/src/bwl/Programm2.java
@@ -19,6 +19,7 @@ import bwl.oo.paket6.*;
 import bwl.oo.paket7.*;
 import bwl.oo.paket8.*;
 import bwl.oo.paket9.*;
+import com.ml.views.Fenster;
 
 public class Programm2 {
