+++ 00_bwl/bwl/src/bwl/Programm3.java
@@ -12,6 +12,7 @@ import com.ml.datenbank.*;
 import bwl.oo.paket1.*;
 import bwl.oo.paket2.*;
 import bwl.oo.paket3.*;
+import com.ml.views.Fenster;
 
 public class Programm3 {
