+++ 00_bwl/bwl/src/bwl/oo/paket3/Angestellter.java
@@ -1,17 +1,22 @@
 package bwl.oo.paket3;
 
-public class Angestellter {
-  
-  public String vorname;
-  public String nachname;
-  
-  
-  public Angestellter(String vorname, String nachname) {
-    super();
-    this.vorname = vorname;
-    this.nachname = nachname;
-  }
-  
-
+public class Angestellter extends Klassen {
+    public Angestellter(){
+        klasse = "Angestellter";
+        instanzVariablenBeschriftung.add("Vorname");
+        instanzVariablenBeschriftung.add("Nachname");
+        instanzVariablenBeschriftung.add("Alter");
+    }
 }
+
+
+
+
+
+
+
+
+
+
+
