+++ 00_bwl/bwl/src/bwl/oo/paket3/Vorlesung.java
@@ -1,16 +1,26 @@
 package bwl.oo.paket3;
 
-import java.util.Vector;
-
-public class Vorlesung {
-  
-  public String titel;
-  public Angestellter professor;
-  
-  public Vorlesung(String titel, Angestellter professor) {
-    super();
-    this.titel = titel;
-    this.professor = professor;
-  }  
+public class Vorlesung extends Klassen {
+    public Vorlesung(){
+        klasse = "Vorlesung";
+        instanzVariablenBeschriftung.add("Name");
+        instanzVariablenBeschriftung.add("Datum");
+    }
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
