--- 00_bwl/bwl/src/bwl/oo/paket3/XAngestellter.java
@@ -1,22 +1,0 @@
-package bwl.oo.paket3;
-
-public class XAngestellter extends XKlassen {
-    public XAngestellter (){
-        klasse = "Angestellter";
-        instanzVariablenBeschriftung.add("Vorname");
-        instanzVariablenBeschriftung.add("Nachname");
-        instanzVariablenBeschriftung.add("Alter");
-    }
-}
-
-
-
-
-
-
-
-
-
-
-
-
