--- 00_bwl/bwl/src/bwl/oo/paket3/XFachschaft.java
@@ -1,23 +1,0 @@
-package bwl.oo.paket3;
-
-public class XFachschaft  extends XKlassen {
-    public XFachschaft (){
-        klasse = "Fachschaft";
-        instanzVariablenBeschriftung.add("Name");
-    }
-
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
