--- 00_bwl/bwl/src/bwl/oo/paket3/XHochschule.java
@@ -1,17 +1,0 @@
-package bwl.oo.paket3;
-
-public class XHochschule extends XKlassen {
-    public XHochschule (){
-        klasse = "Hochschule";
-        instanzVariablenBeschriftung.add("Name");
-    }
-}
-
-
-
-
-
-
-
-
-
