--- 00_bwl/bwl/src/bwl/oo/paket3/XKlassen.java
@@ -1,16 +1,0 @@
-package bwl.oo.paket3;
-
-import java.util.ArrayList;
-import java.util.stream.Collectors;
-
-public abstract class XKlassen {
-    String klasse;
-    ArrayList<String> instanzVariablenBeschriftung = new ArrayList<>();
-    ArrayList<String> instanzVariableWert = new ArrayList<>();
-    public String getInfo(){
-        String info=instanzVariableWert.stream().map(n -> String.valueOf(n))
-                    .collect(Collectors.joining("-", "{", "}"));
-        return info;
-    }
-}
-
