--- 00_bwl/bwl/src/bwl/oo/paket3/XVorlesung.java
@@ -1,26 +1,0 @@
-package bwl.oo.paket3;
-
-public class XVorlesung  extends XKlassen {
-    public XVorlesung (){
-        klasse = "Vorlesung";
-        instanzVariablenBeschriftung.add("Name");
-        instanzVariablenBeschriftung.add("Datum");
-    }
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
